Fix printing using patch from AUR package prior to this
Original source of patch: https://gist.github.com/ShadowKyogre/5249804

More specifically, it was surrounding the arguments with literal quotes and
handled the page numbers in a way that lpr could not handle. Arguments
added to a string list are treated literally and so should not be requoted.

Just make sure that the start page and end page indexing doesn't break 
under windows too since I didn't test that yet.